workspace/backend-amp-core/requirements.txt
fastapi==0.78.0
uvicorn[standard]==0.17.6
sqlalchemy==1.4.37
alembic==1.8.0
pydantic==1.9.0
python-jose==3.3.0
psycopg2-binary==2.9.3